update : Verilator 4.026 Released (add intV, longV, shortV)&header & TCP optimization

+++ utils/c/include/shunt_client_server.h
@@ -120,7 +120,7 @@ typedef struct cs_header_t {
   long   trnx_type;
   long   trnx_id;
   long   data_type;
-  int    n_payloads;
+  long   n_payloads;
 } cs_header;
 
 /*
